+++ src/test/groovy/org/osiam/resources/scim/GroupSpec.groovy
@@ -34,7 +34,8 @@ class GroupSpec extends Specification {
 
         MemberRef memberRef = new MemberRef.Builder().setValue('id').build()
 
-        Group.Builder builder = new Group.Builder('display')
+        Group.Builder builder = new Group.Builder()
+                .setDisplayName('display')
                 .setExternalId('externalId')
                 .setMembers([memberRef] as Set)
 
@@ -50,16 +51,31 @@ class GroupSpec extends Specification {
 
     def 'should be able to add member to group'() {
         when:
-        def group = new Group.Builder('display')
+        def group = new Group.Builder()
+                .setDisplayName('display')
                 .addMember(new MemberRef.Builder().build()).build()
 
         then:
         group.members.size() == 1
     }
 
+    def 'pre-set display name with constructor' () {
+        given:
+        def displayName = 'display name'
+
+        when:
+        def group = new Group.Builder(displayName)
+                .build()
+
+        then:
+        group.displayName == displayName
+    }
+
     def 'should be able to clone a group'() {
         given:
-        def group = new Group.Builder('display').build()
+        def group = new Group.Builder()
+                .setDisplayName('display')
+                .build()
         when:
         def result = new Group.Builder(group).build()
 
@@ -76,9 +92,11 @@ class GroupSpec extends Specification {
         thrown(SCIMDataValidationException)
     }
 
-    def 'the copied group should have the given displayname'() {
+    def 'the copied group should have the given display name'() {
         given:
-        Group oldGroup = new Group.Builder("oldDisplayName").setExternalId("externalId").build()
+        Group oldGroup = new Group.Builder()
+                .setDisplayName('oldDisplayName')
+                .setExternalId('externalId').build()
         String newDisplayName = 'newDisplayName'
         Group newGroup
 
@@ -94,7 +112,8 @@ class GroupSpec extends Specification {
         given:
         MemberRef memberRef = new MemberRef.Builder().setValue('id').build()
 
-        Group.Builder builder = new Group.Builder('display')
+        Group.Builder builder = new Group.Builder()
+                .setDisplayName('display')
                 .setExternalId('externalId')
                 .setMembers([memberRef] as Set)